- Removed hacky way of getting drag and drop to work - Added drag and drop properties to background, lasso, and whiteboard
