+++ webstack/apps/webapp/src/app/pages/board/Board.tsx
@@ -85,7 +85,9 @@ export function BoardPage() {
 
   function handleDragOver(event: DragEvent) {
     const elt = event.target as HTMLElement;
-    if (elt.id !== 'board') {
+    const ids = ['board', 'whiteboard', 'lasso'];
+    if (!ids.includes(elt.id)) {
+      //elt.id !== 'board'
       if (event.dataTransfer) {
         event.dataTransfer.dropEffect = 'none';
       }
